+++ src/main/java/org/torproject/descriptor/impl/DirectoryKeyCertificateImpl.java
@@ -80,6 +80,10 @@ public class DirectoryKeyCertificateImpl extends DescriptorImpl
           crypto.append(line).append(NL);
           break;
         case CRYPTO_END:
+          if (null == crypto) {
+            throw new DescriptorParseException(Key.CRYPTO_END + " before "
+                + Key.CRYPTO_BEGIN);
+          }
           crypto.append(line).append(NL);
           String cryptoString = crypto.toString();
           crypto = null;