+++ src/main/java/org/torproject/descriptor/impl/DirectorySignatureImpl.java
@@ -70,6 +70,10 @@ public class DirectorySignatureImpl implements DirectorySignature {
           crypto.append(line).append(NL);
           break;
         case CRYPTO_END:
+          if (null == crypto) {
+            throw new DescriptorParseException(Key.CRYPTO_END + " before "
+                + Key.CRYPTO_BEGIN);
+          }
           crypto.append(line).append(NL);
           String cryptoString = crypto.toString();
           crypto = null;