+++ src/main/java/org/torproject/descriptor/impl/RelayNetworkStatusImpl.java
@@ -81,6 +81,10 @@ public class RelayNetworkStatusImpl extends NetworkStatusImpl
           crypto.append(line).append(NL);
           break;
         case CRYPTO_END:
+          if (null == crypto) {
+            throw new DescriptorParseException(Key.CRYPTO_END + " before "
+                + Key.CRYPTO_BEGIN);
+          }
           crypto.append(line).append(NL);
           String cryptoString = crypto.toString();
           crypto = null;
@@ -130,6 +134,10 @@ public class RelayNetworkStatusImpl extends NetworkStatusImpl
           crypto.append(line).append(NL);
           break;
         case CRYPTO_END:
+          if (null == crypto) {
+            throw new DescriptorParseException(Key.CRYPTO_END + " before "
+                + Key.CRYPTO_BEGIN);
+          }
           crypto.append(line).append(NL);
           String cryptoString = crypto.toString();
           crypto = null;