+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# Changes in version 2.?.? - 2019-1?-??
+
+
 # Changes in version 2.9.1 - 2019-11-09
 
  * Minor changes